+++ src/main/java/model/Calculator.java
@@ -6,16 +6,16 @@ import java.util.List;
 
 public class Calculator {
     private List<String> separators;
-    private List<Integer> operands;
+    private List<Long> operands;
 
     private Calculator(CalculatorBuilder builder) {
         this.separators = builder.separators;
         this.operands = builder.operands;
     }
 
-    public int plus() {
-        int result = 0;
-        for (Integer operand : operands) {
+    public long plus() {
+        long result = 0L;
+        for (Long operand : operands) {
             if (operand < 0) {
                 throw new IllegalArgumentException("양의 정수만 입력 가능합니다. 잘못된 숫자: " + operand);
             }
@@ -36,20 +36,20 @@ public class Calculator {
         this.separators.add(separator);
     }
 
-    public void addOperand(int operand) {
+    public void addOperand(long operand) {
         this.operands.add(operand);
     }
 
     public static class CalculatorBuilder {
         private List<String> separators = new ArrayList<>();
-        private List<Integer> operands = new ArrayList<>();
+        private List<Long> operands = new ArrayList<>();
 
         public CalculatorBuilder separator(String separator) {
             this.separators.add(separator);
             return this;
         }
 
-        public CalculatorBuilder operand(int operand) {
+        public CalculatorBuilder operand(long operand) {
             this.operands.add(operand);
             return this;
         }
