+++ src/main/java/view/OutView.java
@@ -4,7 +4,8 @@ public class OutView {
     private OutView() {
     }
 
-    public static void outputString(int result) {
+    public static void outputString(long result) {
+        
         System.out.println("결과 : " + result);
     }
 }
